Replace OpenPgpKeyPreference with compat version
