+++ k9mail/src/main/res/xml/account_settings.xml
@@ -416,7 +416,7 @@
             android:key="openpgp_provider"
             android:title="@string/account_settings_crypto_app" />
 
-        <org.openintents.openpgp.util.OpenPgpKeyPreferenceCompat
+        <org.openintents.openpgp.util.OpenPgpKeyPreference
             android:dependency="openpgp_provider"
             android:key="openpgp_key"
             android:title="@string/account_settings_crypto_key" />
